requirements-windows-313.txt
typer==0.12.3
pydantic==2.8.2
httpx==0.27.2
parsel==1.9.1
w3lib==2.2.1
cssselect==1.2.0
tenacity==9.0.0
tqdm==4.66.4
PyYAML>=6.0.2
orjson==3.10.7
numpy==2.3.2
pandas==2.3.2
matplotlib>=3.9.2
lxml
